requirements.txt
tensorflow-gpu==1.13.1
Keras==2.3.1
numpy==1.16.5
imageio==2.6.1
opencv_python_headless==4.1.2.30
matplotlib==3.1.2
albumentations==0.4.3
Pillow==6.2.1
PyYAML==5.2
editdistance==0.5.3